+++ core/data/build.gradle.kts
@@ -9,12 +9,5 @@ setNamespace("core.data")
 
 dependencies {
     implementation(projects.core.dataApi)
-
-    // Network - Ktor
-    implementation(libs.ktor.client.core)
-    implementation(libs.ktor.client.okhttp)
-    implementation(libs.ktor.client.content.negotiation)
-    implementation(libs.ktor.serialization.kotlinx.json)
-    implementation(libs.ktor.client.logging)
     implementation(libs.kotlinx.serialization.json)
 }
