+++ core/remote/build.gradle.kts
@@ -11,6 +11,13 @@ dependencies {
     implementation(projects.core.dataApi)
     implementation(libs.kotlinx.serialization.json)
 
+    // Network - Ktor
+    implementation(libs.ktor.client.core)
+    implementation(libs.ktor.client.okhttp)
+    implementation(libs.ktor.client.content.negotiation)
+    implementation(libs.ktor.serialization.kotlinx.json)
+    implementation(libs.ktor.client.logging)
+
     // Firebase
     implementation(platform(libs.firebase.bom))
     implementation(libs.firebase.database.ktx)
