+++ core/remote/src/main/kotlin/com/teammanduk/adego/core/remote/di/RemoteModule.kt
@@ -3,9 +3,11 @@ package com.teammanduk.adego.core.remote.di
 import com.teammanduk.adego.core.data_api.datasource.LocationDataSource
 import com.teammanduk.adego.core.data_api.datasource.PlaceDataSource
 import com.teammanduk.adego.core.data_api.datasource.RoomDataSource
+import com.teammanduk.adego.core.data_api.datasource.RouteDataSource
 import com.teammanduk.adego.core.remote.datasource.FirebaseRoomDataSource
 import com.teammanduk.adego.core.remote.datasource.FusedLocationDataSource
 import com.teammanduk.adego.core.remote.datasource.TmapPlaceDataSourceImpl
+import com.teammanduk.adego.core.remote.datasource.TmapRouteDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -32,4 +34,10 @@ abstract class RemoteModule {
     abstract fun bindPlaceDataSource(
         tmapPlaceDataSource: TmapPlaceDataSourceImpl
     ): PlaceDataSource
+
+    @Binds
+    @Singleton
+    abstract fun bindRouteDataSource(
+        tmapRouteDataSource: TmapRouteDataSourceImpl
+    ): RouteDataSource
 }
